modify: BaseRandomの生成まわりの整理
